sensor/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
DateTime==4.2
decorator==4.1.2
enum34==1.1.6
functools32==3.2.3.post2
idna==2.6
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-core==4.3.0
nbformat==4.3.0
paho-mqtt==1.3.0
plotly==2.0.14
pytz==2017.2
requests==2.18.4
six==1.10.0
traitlets==4.3.2
urllib3==1.22
zope.interface==4.4.2
